+++ tokio-util/Cargo.toml
@@ -45,6 +45,7 @@ tokio = { version = "0.2.0", path = "../tokio", features = ["full"] }
 tokio-test = { version = "0.2.0", path = "../tokio-test" }
 
 futures = "0.3.0"
+futures-test = "0.3.5"
 
 [package.metadata.docs.rs]
 all-features = true
