+++ tokio-util/src/sync/mod.rs
@@ -1,0 +1,6 @@
+//! Synchronization primitives
+
+mod cancellation_token;
+pub use cancellation_token::{CancellationToken, WaitForCancellationFuture};
+
+mod intrusive_double_linked_list;
