+++ tokio/src/macros/cfg.rs
@@ -354,16 +354,6 @@ macro_rules! cfg_uds {
     }
 }
 
-macro_rules! cfg_unstable {
-    ($($item:item)*) => {
-        $(
-            #[cfg(tokio_unstable)]
-            #[cfg_attr(docsrs, doc(cfg(tokio_unstable)))]
-            $item
-        )*
-    }
-}
-
 macro_rules! cfg_trace {
     ($($item:item)*) => {
         $(
